refactor: rename a few moduels / packages
1. sensetive-permission -> privacy
2. battery-metric -> battery-optim
3. the caliper-gradle-plugin changed the package name to me.xx2bab.caliper.gradle
4. add all caliper modules to sample app subsititution rules
