+++ caliper-gradle-plugin/src/integrationTest/kotlin/me/xx2bab/caliper/CaliperExtensionTest.kt
@@ -1,5 +1,6 @@
 package me.xx2bab.caliper
 
+import me.xx2bab.caliper.gradle.CaliperExtension
 import org.gradle.testfixtures.ProjectBuilder
 import org.hamcrest.MatcherAssert.assertThat
 import org.junit.jupiter.api.Test
