+++ caliper-gradle-plugin/src/integrationTest/kotlin/me/xx2bab/caliper/proxy/ClassProxyTest.kt
@@ -4,8 +4,8 @@ import com.tschuchort.compiletesting.KotlinCompilation
 import com.tschuchort.compiletesting.SourceFile
 import me.xx2bab.caliper.common.ProxiedClass
 import me.xx2bab.caliper.core.ASMManipulator
-import me.xx2bab.caliper.core.CaliperClassVisitor
-import me.xx2bab.caliper.core.ProxyConfig
+import me.xx2bab.caliper.gradle.core.CaliperClassVisitor
+import me.xx2bab.caliper.gradle.core.ProxyConfig
 import me.xx2bab.caliper.tool.invokeMethod
 import org.hamcrest.MatcherAssert
 import org.hamcrest.Matchers
