+++ caliper-gradle-plugin/src/main/kotlin/me/xx2bab/caliper/gradle/core/CaliperClassVisitor.kt
@@ -1,4 +1,4 @@
-package me.xx2bab.caliper.core
+package me.xx2bab.caliper.gradle.core
 
 import org.objectweb.asm.*
