+++ caliper-gradle-plugin/src/main/kotlin/me/xx2bab/caliper/gradle/core/CaliperMethodVisitor.kt
@@ -1,4 +1,4 @@
-package me.xx2bab.caliper.core
+package me.xx2bab.caliper.gradle.core
 
 import me.xx2bab.caliper.anno.ASMOpcodes
 import me.xx2bab.caliper.common.Constants.CALIPER_PACKAGE_FOR_WRAPPER_SPLIT_BY_SLASH
