+++ caliper-runtime-battery-optim/src/main/AndroidManifest.xml
