+++ caliper-runtime-privacy/src/main/AndroidManifest.xml
